Move target related code into it's own directory
This should make it easier to navigate the code base
